# backend/ASTU-backend-group-1/delivery/docs/swagger.yaml
definitions:
...
  infrastructure.Data:
    properties:
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Testing Swagger APIs.
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
  /ai/chat:
    post:
      consumes:
      - application/json
      description: Send a message to the AI model and get a response
      parameters:
      - description: Chat Message
        in: body
        name: message
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat with the AI model
      tags:
      - AI
  /ai/recommend/tags:
    post:
      consumes:
      - application/json
      description: Generate recommended tags based on the provided data
      parameters:
      - description: Input Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/infrastructure.Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recommend tags for the blog
      tags:
      - AI
  /ai/recommend/title:
    post:
      consumes:
      - application/json
      description: Generate a recommended title based on the provided content
      parameters:
      - description: Input Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/infrastructure.Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recommend a title for the blog
      tags:
      - AI
  /ai/refine/{blogId}:
    post:
      consumes:
      - application/json
      description: Refine the content of a blog based on its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refine a blog's content
      tags:
      - blog
  /ai/summarize:
    post:
      consumes:
      - application/json
      description: Generate a summary of the provided content
      parameters:
      - description: Input Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/infrastructure.Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Summarize the provided content
      tags:
      - AI
  /ai/summarize/{blogId}:
    post:
      consumes:
      - application/json
      description: Generate a summary of a blog's content based on its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "406":
          description: Not Acceptable
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Summarize a blog by ID
      tags:
      - blog
...
securityDefinitions:
  JWT:
    in: header
    name: token
    type: apiKey
swagger: "2.0"